Remove the need to clone the canvas
