+++ src/platform_impl/web/stdweb/mod.rs
@@ -5,6 +5,7 @@ mod timeout;
 pub use self::canvas::Canvas;
 pub use self::timeout::Timeout;
 
+use crate::dpi::LogicalSize;
 use crate::platform::web::WindowExtStdweb;
 use crate::window::Window;
 
@@ -12,7 +13,7 @@ use stdweb::js;
 use stdweb::web::event::BeforeUnloadEvent;
 use stdweb::web::window;
 use stdweb::web::IEventTarget;
-use stdweb::web::{document, html_element::CanvasElement};
+use stdweb::web::{document, Element, html_element::{CanvasElement}};
 
 pub fn throw(msg: &str) {
     js! { throw @{msg} }
@@ -29,5 +30,23 @@ pub fn on_unload(mut handler: impl FnMut() + 'static) {
 impl WindowExtStdweb for Window {
     fn canvas(&self) -> CanvasElement {
         self.window.canvas().raw().clone()
+    }
+}
+
+pub fn window_size() -> LogicalSize {
+    let window = window();
+    let width = window.inner_width() as f64;
+    let height = window.inner_height() as f64;
+
+    LogicalSize { width, height }
+}
+
+pub fn is_fullscreen(canvas: &CanvasElement) -> bool {
+    match document().fullscreen_element() {
+        Some(elem) => {
+            let raw: Element = canvas.clone().into();
+            raw == elem
+        }
+        None => false,
     }
 }
