+++ src/platform_impl/web/web_sys/canvas.rs
@@ -9,7 +9,7 @@ use std::rc::Rc;
 
 use wasm_bindgen::{closure::Closure, JsCast};
 use web_sys::{
-    Element, Event, FocusEvent, HtmlCanvasElement, KeyboardEvent, PointerEvent, WheelEvent,
+    Event, FocusEvent, HtmlCanvasElement, KeyboardEvent, PointerEvent, WheelEvent,
 };
 
 pub struct Canvas {
@@ -27,30 +27,6 @@ pub struct Canvas {
     on_mouse_wheel: Option<Closure<dyn FnMut(WheelEvent)>>,
     on_fullscreen_change: Option<Closure<dyn FnMut(Event)>>,
     wants_fullscreen: Rc<RefCell<bool>>,
-    // Used to resize canvas when it enters and exits fullscreen
-    intended_size: Rc<RefCell<LogicalSize>>,
-}
-
-impl Clone for Canvas {
-    fn clone(&self) -> Canvas {
-        Canvas {
-            raw: self.raw.clone(),
-            on_blur: None,
-            on_focus: None,
-            on_keyboard_release: None,
-            on_keyboard_press: None,
-            on_received_character: None,
-            on_cursor_leave: None,
-            on_cursor_enter: None,
-            on_cursor_move: None,
-            on_mouse_release: None,
-            on_mouse_press: None,
-            on_mouse_wheel: None,
-            on_fullscreen_change: None,
-            wants_fullscreen: self.wants_fullscreen.clone(),
-            intended_size: self.intended_size.clone(),
-        }
-    }
 }
 
 impl Drop for Canvas {
@@ -97,10 +73,6 @@ impl Canvas {
             on_mouse_wheel: None,
             on_fullscreen_change: None,
             wants_fullscreen: Rc::new(RefCell::new(false)),
-            intended_size: Rc::new(RefCell::new(LogicalSize {
-                width: 0.0,
-                height: 0.0,
-            })),
         })
     }
 
@@ -124,33 +96,9 @@ impl Canvas {
         self.raw.height() as f64
     }
 
-    pub fn set_size(&self, size: LogicalSize, intended: bool) {
+    pub fn set_size(&self, size: LogicalSize) {
         self.raw.set_width(size.width as u32);
         self.raw.set_height(size.height as u32);
-
-        if intended {
-            *self.intended_size.borrow_mut() = size;
-        }
-    }
-
-    pub fn intended_size(&self) -> LogicalSize {
-        *self.intended_size.borrow()
-    }
-
-    pub fn window_size(&self) -> LogicalSize {
-        let window = web_sys::window().expect("Failed to obtain window");
-        let width = window
-            .inner_width()
-            .expect("Failed to get width")
-            .as_f64()
-            .expect("Failed to get width as f64");
-        let height = window
-            .inner_height()
-            .expect("Failed to get height")
-            .as_f64()
-            .expect("Failed to get height as f64");
-
-        LogicalSize { width, height }
     }
 
     pub fn raw(&self) -> &HtmlCanvasElement {
@@ -352,15 +300,6 @@ impl Canvas {
     }
 
     pub fn is_fullscreen(&self) -> bool {
-        let window = web_sys::window().expect("Failed to obtain window");
-        let document = window.document().expect("Failed to obtain document");
-
-        match document.fullscreen_element() {
-            Some(elem) => {
-                let raw: Element = self.raw.clone().into();
-                raw == elem
-            }
-            None => false,
-        }
+        super::is_fullscreen(&self.raw)
     }
 }
