+++ src/platform_impl/web/web_sys/mod.rs
@@ -5,10 +5,11 @@ mod timeout;
 pub use self::canvas::Canvas;
 pub use self::timeout::Timeout;
 
+use crate::dpi::LogicalSize;
 use crate::platform::web::WindowExtWebSys;
 use crate::window::Window;
 use wasm_bindgen::{closure::Closure, JsCast};
-use web_sys::{BeforeUnloadEvent, HtmlCanvasElement};
+use web_sys::{BeforeUnloadEvent, Element, HtmlCanvasElement, window};
 
 pub fn throw(msg: &str) {
     wasm_bindgen::throw_str(msg);
@@ -36,5 +37,34 @@ pub fn on_unload(mut handler: impl FnMut() + 'static) {
 impl WindowExtWebSys for Window {
     fn canvas(&self) -> HtmlCanvasElement {
         self.window.canvas().raw().clone()
+    }
+}
+
+pub fn window_size() -> LogicalSize {
+    let window = web_sys::window().expect("Failed to obtain window");
+    let width = window
+        .inner_width()
+        .expect("Failed to get width")
+        .as_f64()
+        .expect("Failed to get width as f64");
+    let height = window
+        .inner_height()
+        .expect("Failed to get height")
+        .as_f64()
+        .expect("Failed to get height as f64");
+
+    LogicalSize { width, height }
+}
+
+pub fn is_fullscreen(canvas: &HtmlCanvasElement) -> bool {
+    let window = window().expect("Failed to obtain window");
+    let document = window.document().expect("Failed to obtain document");
+
+    match document.fullscreen_element() {
+        Some(elem) => {
+            let raw: Element = canvas.clone().into();
+            raw == elem
+        }
+        None => false,
     }
 }
